disable opening link in web when working with cody app
