+++ client/jetbrains/CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 - Use smaller Cody logo in toolbar and editor context menu [#54481](https://github.com/sourcegraph/sourcegraph/pull/54481)
+- Sourcegraph link sharing and opening file in browser actions are disabled when working with Cody app [#54473](https://github.com/sourcegraph/sourcegraph/pull/54473)
 
 ### Deprecated
