+++ client/jetbrains/src/main/java/com/sourcegraph/website/CopyAction.java
@@ -3,11 +3,8 @@ package com.sourcegraph.website;
 import com.intellij.notification.Notification;
 import com.intellij.notification.NotificationType;
 import com.intellij.notification.Notifications;
-import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.ide.CopyPasteManager;
 import com.intellij.openapi.project.Project;
-import com.sourcegraph.config.ConfigUtil;
-import com.sourcegraph.config.SettingsComponent;
 import java.awt.datatransfer.StringSelection;
 import org.jetbrains.annotations.NotNull;
 
@@ -28,16 +25,5 @@ public class CopyAction extends FileActionBase {
             "File URL copied to clipboard: " + urlWithoutUtm,
             NotificationType.INFORMATION);
     Notifications.Bus.notify(notification);
-  }
-
-  @Override
-  public void update(@NotNull AnActionEvent e) {
-    Project project = e.getProject();
-    if (project == null) {
-      return;
-    }
-    e.getPresentation()
-        .setEnabled(
-            ConfigUtil.getInstanceType(project) != SettingsComponent.InstanceType.LOCAL_APP);
   }
 }
